moviendo la consultas a los models, endpoint para crear rifa y endpoint para finalizar una rifa
