+++ BackEnd/routers/admin_router.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, Depends
 from schemas.Rifa_schema import Rifa
-from models.rifa_service import RifaService
+from models.rifa_model import RifaService
 from db.conexion import session_dependency
+from fastapi.responses import JSONResponse
 
 admin_router = APIRouter()
 
@@ -13,6 +14,11 @@ def loginAdmin():
 
 
 # Ednpoint para crear rifa
-@admin_router.post("/rifas")
+@admin_router.post("/rifa")
 def crear_rifa(rifa: Rifa, session: session_dependency):
     return RifaService(session=session).crear_rifa(rifa=rifa)
+
+
+@admin_router.put("rifa/{id}")
+def finalizar_rifa(id: int, session: session_dependency):
+    return RifaService(session=session).finalizar_rifa(id=id)
